refactor(view): move Input to view package
