+++ src/test/java/racingcar/implementtest/InputTest.java
@@ -7,7 +7,7 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import racingcar.Input;
+import racingcar.view.Input;
 
 public class InputTest {
     private Input input;
